version 1.4.0

+++ contrib/cirrus/integration_test.sh
@@ -5,11 +5,18 @@ source $(dirname $0)/lib.sh
 
 req_env_var GOSRC SCRIPT_BASE OS_RELEASE_ID OS_RELEASE_VER CONTAINER_RUNTIME
 
+# Our name must be of the form xxxx_test or xxxx_test.sh, where xxxx is
+# the test suite to run; currently (2019-05) the only option is 'integration'
+# but pr2947 intends to add 'system'.
+TESTSUITE=$(expr $(basename $0) : '\(.*\)_test')
+if [[ -z $TESTSUITE ]]; then
+    die 1 "Script name is not of the form xxxx_test.sh"
+fi
+
 cd "$GOSRC"
 
 if [[ "$SPECIALMODE" == "in_podman" ]]
 then
-    set -x
     ${CONTAINER_RUNTIME} run --rm --privileged --net=host \
         -v $GOSRC:$GOSRC:Z \
         --workdir $GOSRC \
@@ -20,41 +27,30 @@ then
         -e "CONMON_BINARY=/usr/libexec/podman/conmon" \
         -e "DIST=$OS_RELEASE_ID" \
         -e "CONTAINER_RUNTIME=$CONTAINER_RUNTIME" \
-        ${OS_RELEASE_ID}podmanbuild bash $GOSRC/$SCRIPT_BASE/container_test.sh -b -i -t -n
+        $IN_PODMAN_IMAGE bash $GOSRC/$SCRIPT_BASE/container_test.sh -b -i -t
 
     exit $?
 elif [[ "$SPECIALMODE" == "rootless" ]]
 then
     req_env_var ROOTLESS_USER
-    set -x
-    ssh $ROOTLESS_USER@localhost \
+
+    if [[ "$USER" == "$ROOTLESS_USER" ]]
+    then
+        $GOSRC/$SCRIPT_BASE/rootless_test.sh ${TESTSUITE}
+    else
+        ssh $ROOTLESS_USER@localhost \
                 -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no -o CheckHostIP=no \
-                $GOSRC/$SCRIPT_BASE/rootless_test.sh
-    exit $?
+                $GOSRC/$SCRIPT_BASE/rootless_test.sh ${TESTSUITE}
+    fi
 else
-    set -x
     make
     make install PREFIX=/usr ETCDIR=/etc
     make test-binaries
-    make install.tools
-    clean_env
-
-    case "${OS_RELEASE_ID}-${OS_RELEASE_VER}" in
-        ubuntu-18) ;;
-        fedora-29) ;&  # Continue to the next item
-        fedora-28) ;&
-        centos-7) ;&
-        rhel-7)
-            make podman-remote
-            install bin/podman-remote /usr/bin
-            ;;
-        *) bad_os_id_ver ;;
-    esac
     if [[ "$TEST_REMOTE_CLIENT" == "true" ]]
     then
-        make remoteintegration
+        make remote${TESTSUITE}
     else
-        make localintegration
+        make local${TESTSUITE}
     fi
     exit $?
 fi
